backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-multipart==0.0.9
sqlalchemy[asyncio]==2.0.30
aiosqlite==0.20.0
asyncpg==0.29.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.3
scikit-learn==1.5.0
numpy==1.26.4
pandas==2.2.2
joblib==1.4.2
pydantic==2.7.1
pydantic[email]==2.7.1
email-validator==2.1.1
python-dotenv==1.0.1
httpx==0.27.0
